Todas la Mejoras

+++ C#_Aplication/Agilize/SignUp.cs
@@ -14,13 +14,18 @@ using System.Windows.Forms;
 
 namespace Agilize
 {
+    
     public partial class SignUp : Form
     {
+        
         Users newUser = new Users();
         String pathToProjectFiles;
-        
-        String encryptingKey = "f83jsd74jdue0qnd";// Letras aleatoreas completamente
+        bool validUser = true;
+        String encryptingKey = "f83jsd74jdue0qnd";// Clave para el ecryptado de blowFish, son letras y numeros aleatoreos
 
+        /// <summary>
+        /// Contructor del form, recibe el path donde estan los archivos del programa.
+        /// </summary>
         public SignUp(String pathToProjectFiles)
         {
             InitializeComponent();
@@ -28,12 +33,18 @@ namespace Agilize
             this.pathToProjectFiles = pathToProjectFiles;
         }
 
+        /// <summary>
+        /// Settea todo el apartado visual del form
+        /// </summary>
         private void SetAll()
         {
             LoginLbl.LinkBehavior = System.Windows.Forms.LinkBehavior.NeverUnderline;
             RedondearBoton(signUpBtn);
         }
 
+        /// <summary>
+        /// Redondea los botones
+        /// </summary>
         private void RedondearBoton(System.Windows.Forms.Button btn)
         {
             var radio = 15;
@@ -48,6 +59,10 @@ namespace Agilize
             btn.Region = new Region(path);
         }
 
+        /// <summary>
+        /// Comprueba que el usuario que se intenta registrar sea valido (No tenga información vacia y comprueba que no exista ya),
+        /// lo crea y lo añade al Json de Usuarios.
+        /// </summary>
         private void signUpBtn_Click(object sender, EventArgs e)
         {
             newUser.password = EncryptPassword(newUser.password);
@@ -67,15 +82,32 @@ namespace Agilize
                     usersList = System.Text.Json.JsonSerializer.Deserialize<List<Users>>(jsonContent);
                 }
 
-                // Agregar el nuevo usuario a la lista
-                usersList.Add(newUser);
+                foreach (Users user in usersList)
+                {
+                    if (newUser.nickname.Equals(user.nickname))
+                    {
+                        MessageBox.Show("El Nickname de usuario ya existe, por favor usa otro", "Éxito", MessageBoxButtons.OK, MessageBoxIcon.Information);
+                        NicknameTxtBox.Text = "Nickname";
+                        validUser = false;
+                        break;
+                    }
+                }
+                if (validUser)
+                {
+                    // Agregar el nuevo usuario a la lista
+                    usersList.Add(newUser);
 
-                string newJsonContent = System.Text.Json.JsonSerializer.Serialize(usersList, new JsonSerializerOptions { WriteIndented = true });
-                File.WriteAllText(pathToProjectFiles, newJsonContent);
+                    string newJsonContent = System.Text.Json.JsonSerializer.Serialize(usersList, new JsonSerializerOptions { WriteIndented = true });
+                    File.WriteAllText(pathToProjectFiles, newJsonContent);
 
-                MessageBox.Show("Usuario registrado con éxito.", "Éxito", MessageBoxButtons.OK, MessageBoxIcon.Information);
+                    MessageBox.Show("Usuario registrado con éxito.", "Éxito", MessageBoxButtons.OK, MessageBoxIcon.Information);
 
-                this.Close();
+                    this.Close();
+                }
+                else
+                {
+                    validUser = true;
+                }
             }
             else
             {
@@ -83,26 +115,37 @@ namespace Agilize
             }
         }
 
+        /// <summary>
+        /// Cierra el formulario para volver al form de Login
+        /// </summary>
         private void LoginLbl_LinkClicked(object sender, LinkLabelLinkClickedEventArgs e)
         {
             this.Close();
         }
 
+        /// <summary>
+        /// Comprueba que el text box tenga como nombre Name, si es así lo borra para que el usuario pueda escribir.
+        /// Comprueba que el texto no sea Name para que pueda funcionar como Hint.
+        /// </summary>
         private void nameTxtBox_Enter(object sender, EventArgs e)
         {
             if (nameTxtBox.Text == "Name")
             {
                 nameTxtBox.Text = "";
-                nameTxtBox.ForeColor = SystemColors.WindowText; // Cambiar a color normal
+                nameTxtBox.ForeColor = SystemColors.WindowText;
             }
         }
 
+        /// <summary>
+        /// Guarda el nombre del nuevo usuario, sino deja el texto de Name para indicar que se ha de poner el nombre
+        /// funciona como un hint.
+        /// </summary>
         private void nameTxtBox_Leave(object sender, EventArgs e)
         {
             if (string.IsNullOrWhiteSpace(nameTxtBox.Text))
             {
                 nameTxtBox.Text = "Name";
-                nameTxtBox.ForeColor = SystemColors.GrayText; // Cambiar a color gris
+                nameTxtBox.ForeColor = SystemColors.GrayText;
             }
             else
             {
@@ -110,21 +153,30 @@ namespace Agilize
             }
         }
 
+
+        /// <summary>
+        /// Comprueba que el text box tenga como nombre Surnames, si es así lo borra para que el usuario pueda escribir.
+        /// Comprueba que el texto no sea Surnames para que pueda funcionar como Hint.
+        /// </summary>
         private void surnamesTxtBox_Enter(object sender, EventArgs e)
         {
             if (surnamesTxtBox.Text == "Surnames")
             {
                 surnamesTxtBox.Text = "";
-                surnamesTxtBox.ForeColor = SystemColors.WindowText; // Cambiar a color normal
+                surnamesTxtBox.ForeColor = SystemColors.WindowText;
             }
         }
 
+        /// <summary>
+        /// Guarda el apellido del nuevo usuario, sino deja el texto de Surnames para indicar que se ha de poner el apellido
+        /// funciona como un hint.
+        /// </summary>
         private void surnamesTxtBox_Leave(object sender, EventArgs e)
         {
             if (string.IsNullOrWhiteSpace(surnamesTxtBox.Text))
             {
                 surnamesTxtBox.Text = "Surnames";
-                surnamesTxtBox.ForeColor = SystemColors.GrayText; // Cambiar a color gris
+                surnamesTxtBox.ForeColor = SystemColors.GrayText;
             }
             else
             {
@@ -132,21 +184,30 @@ namespace Agilize
             }
         }
 
+
+        /// <summary>
+        /// Comprueba que el text box tenga como nombre Email, si es así lo borra para que el usuario pueda escribir.
+        /// Comprueba que el texto no sea Email para que pueda funcionar como Hint.
+        /// </summary>
         private void mailTxtBox_Enter(object sender, EventArgs e)
         {
             if (mailTxtBox.Text == "Email")
             {
                 mailTxtBox.Text = "";
-                mailTxtBox.ForeColor = SystemColors.WindowText; // Cambiar a color normal
+                mailTxtBox.ForeColor = SystemColors.WindowText;
             }
         }
 
+        /// <summary>
+        /// Guarda el mail del nuevo usuario, sino deja el texto de Email para indicar que se ha de poner el mail
+        /// funciona como un hint.
+        /// </summary>
         private void mailTxtBox_Leave(object sender, EventArgs e)
         {
             if (string.IsNullOrWhiteSpace(mailTxtBox.Text))
             {
                 mailTxtBox.Text = "Email";
-                mailTxtBox.ForeColor = SystemColors.GrayText; // Cambiar a color gris
+                mailTxtBox.ForeColor = SystemColors.GrayText;
             }
             else
             {
@@ -154,21 +215,30 @@ namespace Agilize
             }
         }
 
+
+        /// <summary>
+        /// Comprueba que el text box tenga como nombre Nickname, si es así lo borra para que el usuario pueda escribir.
+        /// Comprueba que el texto no sea Nickname para que pueda funcionar como Hint.
+        /// </summary>
         private void NicknameTxtBox_Enter(object sender, EventArgs e)
         {
             if (NicknameTxtBox.Text == "Nickname")
             {
                 NicknameTxtBox.Text = "";
-                NicknameTxtBox.ForeColor = SystemColors.WindowText; // Cambiar a color normal
+                NicknameTxtBox.ForeColor = SystemColors.WindowText;
             }
         }
 
+        /// <summary>
+        /// Guarda el nickname del nuevo usuario, sino deja el texto de Nickname para indicar que se ha de poner el nickname
+        /// funciona como un hint.
+        /// </summary>
         private void NicknameTxtBox_Leave(object sender, EventArgs e)
         {
             if (string.IsNullOrWhiteSpace(NicknameTxtBox.Text))
             {
                 NicknameTxtBox.Text = "Nickname";
-                NicknameTxtBox.ForeColor = SystemColors.GrayText; // Cambiar a color gris
+                NicknameTxtBox.ForeColor = SystemColors.GrayText;
             }
             else
             {
@@ -176,21 +246,30 @@ namespace Agilize
             }
         }
 
+
+        /// <summary>
+        /// Comprueba que el text box tenga como nombre Password, si es así lo borra para que el usuario pueda escribir.
+        /// Comprueba que el texto no sea Password para que pueda funcionar como Hint.
+        /// </summary>
         private void PaswordTxtBox_Enter(object sender, EventArgs e)
         {
             if (PaswordTxtBox.Text == "Password")
             {
                 PaswordTxtBox.Text = "";
-                PaswordTxtBox.ForeColor = SystemColors.WindowText; // Cambiar a color normal
+                PaswordTxtBox.ForeColor = SystemColors.WindowText;
             }
         }
 
+        /// <summary>
+        /// Guarda la contraseña del nuevo usuario, sino deja el texto de Password para indicar que se ha de poner la contraseña
+        /// funciona como un hint.
+        /// </summary>
         private void PaswordTxtBox_Leave(object sender, EventArgs e)
         {
             if (string.IsNullOrWhiteSpace(PaswordTxtBox.Text))
             {
                 PaswordTxtBox.Text = "Password";
-                PaswordTxtBox.ForeColor = SystemColors.GrayText; // Cambiar a color gris
+                PaswordTxtBox.ForeColor = SystemColors.GrayText;
             }
             else
             {
@@ -198,6 +277,9 @@ namespace Agilize
             }
         }
 
+        /// <summary>
+        /// Comprueba que los datos del usuario no sean un valor en nulo o que sean solamente un espacio en blanco.
+        /// </summary>
         private bool ValidateNewUser()
         {
             return !string.IsNullOrWhiteSpace(newUser.name) &&
@@ -207,6 +289,9 @@ namespace Agilize
                    !string.IsNullOrWhiteSpace(newUser.password);
         }
 
+        /// <summary>
+        /// Encripta la contraseña con el metodo BlowFish usando una clave, y devuelve un string con el password encryptado
+        /// </summary>
         public string EncryptPassword(string pswd)
         {
             var engine = new BlowfishEngine();
